ESP-IDF Versions
This is an early version of ESP-IDF based SHTC3 sensors. No work has
been done on the E-Ink display yet, but this has a semi-functional I2C
component reading from the sensor. Some basic error handling is here and
it uses the example Wifi provisioning from esp. Other code is cribbed
form Adafruit and this will be cleaned up to get the licenses right