Kubernetes automation cleanup
This is still somewhat WIP as I was trying to get TimeScaleDB to run
in cluster. I'm likely going to give up on that because running a DB
inside a Kube cluster is just a bad plan anyways. I'm spending more time
futzing with getting the DB in there than I am automating the schema
definition and getting data written to the thing.