MQTT, Kubernets and straight C (#4)
* WIP: Drop C++, add some wip features for MQTT

Trying to make the MQTT client slightly better, auto-disconnect and
eventually expo-backoff. Need to configure some kind of ring buffer for
sensor measurements which mqtt can pluck from.

* Multi-threaded MQTT & Sensor reading

* Sensors now have their own thread to grab and queue up data.
* MQTT Client waits for notifications to ask registered handlers to
  return JSON for their readings.
* MQTT has beginnings of a client watchdog to backoff server connection
  attempts.

* Exponential Backoff for MQTT connections

* WIP Backend Updates

Converting several charts to helm3 for kubernetes. Some of this is
useful others it was a learning experience. Got several things working
though:
* private docker repo
* traefik routing of straight TCP on a port

Going to blast the vagrant host and respin to see if my automation works
now

* JSON Formatting of messages

Started defining a structure for sensor data so that it can post to MQTT
and then migrate into timescaledb. Still iterating on the structure, but
it should be threadsafe and handle local caching of the data (short of
deep sleep) for a bit of time.

* Move kubernetes files around

* WIP Kubernetes standup automation validated

* Kubernetes extneral-dns support... kinda.

Reworked dockerhub and mqtt to not use Traefik since it won't register
with external-dns (https://github.com/traefik/traefik/issues/4655).

Additionally this setup is rather delicate.

* Opensense doesn't really support using Bind as a RFC2136 server, so
  I'm reusing a key which _SHOULD_ be autorotated
  (https://github.com/opnsense/plugins/issues/2196) but isn't.
* I'm manually editing the named.conf file to push my zone into the
  `dynamic` directory so that bind can write it's jnl file. That's just
  a reboot away from not working again :fire: :skull: :fire:

* Publish directly, and use DNS rather than IP's

* Kubernetes automation cleanup

This is still somewhat WIP as I was trying to get TimeScaleDB to run
in cluster. I'm likely going to give up on that because running a DB
inside a Kube cluster is just a bad plan anyways. I'm spending more time
futzing with getting the DB in there than I am automating the schema
definition and getting data written to the thing.